+++ lib/framework.rb
@@ -1,0 +1,5 @@
+require "framework/abilities"
+require "framework/cards"
+
+module Framework
+end
