+++ lib/framework/abilities.rb
@@ -1,0 +1,7 @@
+require "framework/abilities/definition"
+require "framework/abilities/dsl"
+
+module Framework
+  module Abilities
+  end
+end
