+++ lib/framework/cards.rb
@@ -1,0 +1,7 @@
+require "framework/cards/definition"
+require "framework/cards/dsl"
+
+module Framework
+  module Cards
+  end
+end
