+++ lib/realms.rb
@@ -1,6 +1,7 @@
 require "realms/version"
 require "active_support/all"
 require "brainguy"
+require "framework"
 require "realms/game"
 
 module Realms
